\name{animaker-package}
\alias{animaker-package}
\alias{animaker}
\docType{package}
\title{
  Generating Animation Timelines
}
\description{
  Functions for describing animation timelines, plus functions for
  visualising the timelines and generating timing schemes.
}
\author{
  Paul Murrell and Simon Potter.
}
