\name{durn}
\alias{durn}
\alias{durn<-}
\title{
  Getting and setting the duration of an animation
}
\description{
  Simple methods for accessing and manipulating the duration of an
  animation.
}
\usage{
durn(x)

durn(x) <- value
}
\arguments{
  \item{x}{
    An animation object.
  }
  \item{value}{
    The new animation duration.
  }
}
\value{
  None.
}
\author{
  Paul Murrell.
}
