\name{start}
\alias{start}
\alias{start<-}
\title{
  Getting and setting the starting time of an animation
}
\description{
  Simple methods for accessing and manipulating the starting time of an
  animation.
}
\usage{
start(x)

start(x) <- value
}
\arguments{
  \item{x}{
    An animation object.
  }
  \item{value}{
    The new animation starting time.
  }
}
\value{
  None.
}
\author{
  Paul Murrell.
}
